create a font release zip file
The fonts repository has a very large number of individual PF files. This makes it so any script that wishes to download all fonts will take a very long time to do so with tools like wget and curl. This change will enable the updater script to very efficiently grab all fonts and quickly unzip them on the MiSTer locally.